+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### 🚜 Refactor
 
+- *(config)* [**breaking**] Update OnDestinationDirNotFound case
 - *(pre-commit)* Update pre-commit hook versions and add conventional commit hook
 - *(core)* Update ChoiceSelector to be more readable
 - *(uv)* Migrate from poetry to uv